Remove transitions for now
If you want to add transitions, the containers should be IStateful so we can discern their state for later use (because I switched to Show/Hide I can use IsPresent for now). We should probably look at moving the VisibilityState portion of OverlayContainer into a new StatefulContainer class or similar, so it can be used in situations like this.