Fix lead in time.
The issue was that hitobjects are very sensitive to the exact clock
time when their loadComplete is called. Before, PlayerInputManager
would override the clock with its own framed ManualClock, but it did
not set the time and ProcessFrame immediately after init (it only does
this in Update), thus when it came to the HitObject being loaded, it
had a Clock.Current of 0. The end result was sliders that look shit.

https://streamable.com/d9e0i