Remove StandardHUDOverlay
This class hierarchy was originally when we had different HUDs for rulesets, but this is no longer the case. Skins will be able to build off the base (new) HUDOverlay without issue, so there is no real need for the abstract class for now.

We can add it back if a need for it arises; for the moment I'd like to avoid as many Standard* classes as possible.