Add back osu.Desktop project
While it was nice having less projects to deal with, it meant hacks to get around circular dependencies between rulesets and osu.Game. For now, let's just restore the osu.Desktop project and call it a day.

This allows for the project to build again with default visual studio settings.