Make Auto try to hold cursor at end of sliders
Turns out if you start moving the cursor away from the slider, even if
you hold the mouse button down for KEY_UP_DELAY after the end of the
slider, the slider will flash for some reason.

In this fix I do not change the button pressing behaviour, only cursor
movement, but it fixes the slider flashes. Intuitively, whether you move
your mouse a little at the end of the slider or not should not cause the slider
to flash, but it does.