Updated UserStatus.cs to make more sense(UNTESTED)
Untested code, on my ipad in school so i can’t compile the code or check for refrences to stuff

Modding, playing multiplayer/singleplayer, Busy, etc are all done when user is online so they may just be a class of type UserStatusOnline(used to be UserStatusAvailable but since available and online are basically the same i just made available be online). I don’t know how PM’s are handled but i assume that the client recieves the PM and then decides what to do based on your status. By doing this you can interupt the user with the message if typeof(status) != typeof(UserStatusBusy), making multiplaying, solo game, etc not interupt the user, but when typeof(status) == typeof(UserStatusBusy) show a less intrusive message box to not interupt the user