Made it so that time display adjusts its self based on current time.
When there is < 10 minutes, timer displays m:ss, when it is above 10 minutes, it displays mm:ss, and so on. This also happens with the hours and time remaining.
Can't seem to find a way to combine this pull request with the one I posted before, so I am making a new one. I'll reference this one in the orignal post.
This was suggested by swoolcock.
#2475