Lets fix high bpm! (220 BPM+ Buff)
Hello!

I've been wanting to propose this for a while and I've been trying to come up with a good way to build up data but after several nights of trying to convince a friend to help me I decided to compile some small results and ask others (Reddit, w/e) to assist me.

It is widely accepted that high BPM streams are currently not accounted for / balanced properly in the current algorithm. What I am proposing is a buff to all streams that exceed 220 bpm. What I am doing in my edit to the code is checking the MS between the previous note and the one we're currently looking at. If it is below `68` (220 BPM 1/4th), then we start scaling it linearly via the following formula. 

`speedBonus = 68 / current.DeltaTime;`

Which provides a bonus of 1.09x for 240 BPM, 1.18x for 260 BPM, 1.36x for 300 BPM, etc.

Here is an example of this code in action with some values for SS's on several maps:

http://puu.sh/AnpoZ/35b3c81c3c.png

I am willing to collect more data, but my coding knowledge is very limited so doing it by hand is very tedious and not effective at all. If there are people who have the necessary skills and would like to assist, please do so using this PR thread so we can build better statistics and understand how this change will effect much more of the game. The only way this change can be implemented is if we have proper data to show the effects of this change, so lets get to it!

Thank you for reading