Refactor LegacyDecoder parsing logic
When calling Decoder<T>.GetDecoder(), the method examines the first line
of the stream being decoded for a magic string and uses it to return
a correct decoder type. Because it operates on the stream directly, that
first line is consumed - the actual decoder will read from the second
line onwards.

It is however possible that this magic stream will not be present in the
first line of the file, and will instead actually be a part of the
content to be decoded (most likely, a section header). In such a case
it is desired that this line is not discarded, and instead parsed
the same way normal lines are.

Rewinding the StreamReader is not correct due to encoding issues.
If the file being parsed contains a byte sequence indicating encoding,
like the UTF-8 byte-order mark, it will be read with the first line
and break parsing.

To bypass this roadblock, extract current configuration section
to a protected field, and extract logic of parsing a single line
to a protected method. This will allow GetDecoder() to fall back to
the normal line parsing path, when a file without a header is
encountered.