Construct decoded object inside base Decoder
Up to this point derived Decoder implementations operated on objects
which were passed to them by Decoder<TOutput>.Decode(). This makes
things problematic if there is a need to parse a line that can't be
recovered later inside of GetDecoder().

Refactor base Decoder to contain an Output property with the decoded
object instance, which is lazily constructed on-demand using the
implementation of CreateTemplateObject().

This allows to remove fields in LegacyBeatmapDecoder and
LegacyStoryboardDecoder which functionally had the same usage.