Implement fallback decoder registration
After the preparatory refactorings of Decoders, it is now possible to
introduce registration of fallback decoders that won't drop input
supplied in the first line of the file.

A fallback decoder is used when the magic in the first line of the file
does not match any of the other known decoders. In such a case,
the fallback decoder is constructed and passed the first line of input
consumed from the stream for potential reparsing.

There can be only one fallback decoder per type; a second attempt of
registering a fallback will result in an exception to avoid bugs.