Add explicit app manifest to desktop project
After the .NET Core bump to version 3.0 in the 2019.1011.0 release,
reports popped up of the game not starting any more on some computers
using Intel graphics cards (HD 3000 in particular).

After investigation the auto-generated application manifest changed in
.NET Core 3.0. In particular this seems to be a root cause for the
failed start-ups on Intel cards, due to a Windows version compatibility
section appearing. The section in turn affects some WinAPI calls like
GetVersionEx, which will return major version 10 instead of 6 if
compatibility with Windows 10 is declared. This combined with a broken
check in the Intel OpenGL driver caused the crashes.

To resolve this without having to patch binaries, add an explicit
application manifest to the desktop project with the compatibility
section removed.