Fix incorrect fallback logic
The recently-modified skin texture fallback logic was very subtly
incorrect. If at the end of the first loop no texture was found, it
would be checked for null to avoid setting scale adjust on a null
texture, but then returned anyway, bypassing the fallback logic for
subsequent possible paths entirely.

Invert the check and explicitly continue to the next fallback path
if neither a 2x, nor 1x texture with the given name is found in the
store.