Add support for LayeredHitSounds skin option
The LayeredHitSounds option if set to 1 will force hitnormal to be
played along all the other hitsounds on each hitobject.

This option is disabled in osu:mania. In order to implement this
behaviour, an AllowsLayeredHitSounds attribute was added to the
ConvertHitObjectParser so that it knows if it has to artificially add
the hitnormal sample or not. LegacySkin can then verify if a hitsample
was artificially added thanks to HitSampleInfo's IsLayered attribute. It
can then decide to return the hitsample or not based on the value of
LayeredHitSounds.