Increase backwards seek magnitude when the track is running
This matches osu-stable. When the track is running, seeking backwards
(against the flow) is harder than seeking forwards. Adding a mutliplier
makes it feel much better.

Note that this is additive not multiplicative because for larger seeks
the (where `amount` > 1) we don't want to jump an insanely huge amount -
just offset the seek slightly to account for playing audio.