Don't play flair animations / sounds when watching autoplay or viewing a result directly
I think this is a pretty good place to be for now. The flair will play
if you just watched a play (local, replay or spectator) but will not
play if you are coming from song select (viewing a replay's result
screen from the leaderboard) or in the case of autoplay.

Closes #10762.