Improve scroll tracking of selected item
As pointed out in #10934, during searching/filtering, the scroll
position can get out of sync with the current item, causing massive
visual noise while it catches up.

This PR changes the way we track the active item, by keeping a reference
to the active panel rather than its position, allowing correctly
tracking the vertical transform as panels move around each other.

Note that this doesn't track 1:1 (likely due to an off-by-one update?)
but this is actually for the best, and leaves a bit of play in the
animation. I tried making it track perfectly and it doesn't look or feel
as good.

Closes #10934.