Reference CatchHitObject from IHasCatchObjectState again.
The main motivation for removing HitObject reference was that its properties could change after the object is caught in the editor.
It turns out, however, in the current state of PR, HitObject properties being changed after caught doesn't cause issues. The way StartTime currently used doesn't matter StartTime as long as it is after the current time. And the type of HitObject cannot be changed.
I'll make it references to a HitObject for now to make the PR simpler.