Add CustomDifficultySetting
Previously, handling of the custom difficulty values in the Difficulty
Adjust mod was quite awkward, as the mod attempted to track both the
base difficulty and user adjustments in a single bindable property,
which could expectedly fail in some scenarios.

To alleviate that, add a `CustomDifficultySetting` more geared to this
use case, that allows to switch a bindable between a "base", immutable
mode, and a "custom", mutable mode. The former will be used when there
are no user difficulty adjustments applied, while the latter shall be
used when the user enables adjusting a particular difficulty aspect.