Modification by group, remove difficulty speed cap
For every group with a timing point, an associated difficulty point is modified/added. For every group with no timing point but a difficulty point, remove the difficulty point. The second case comes in use as 'Giving every difficulty point a speed multipler of 1' would interfere with difficulty points that counteract timing points.