Replace LimitedCapacityStack with a regular Stack
There is no reason we should be limiting skills to knowing only the previous 2 objects before the current one. This originally existed as an angle implementation detail of the original pp+ codebase which made its way here, but didn't get used in the same way.

If we want to get only the previous 2 objects we can simply use linq. ie. Previous.Take(2)